pyproject.toml
[project]
name = "hl1-audio-catalog"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "mutagen>=1.47.0",
]
